Clean up specs
