+++ spec/segment_spec.rb
@@ -4,31 +4,23 @@ RSpec.describe Edir::Segment do
   describe "#to_h" do
     it "returns the segment in the expected hash representation" do
       elements = [
-        "ISA", "*",
-        "00", "*",
-        "00", "*",
+        "B2", "*",
         "", "*",
-        "09", "*",
-        "005070479ff", "*",
-        "ZZ", "*",
-        "U", "*",
-        "0", "*",
-        "T", "*",
-        "^"
+        "SCAC", "*",
+        "", "*",
+        "124592j3", "*",
+        "", "*",
+        "CC"
       ]
       segment = Edir::Segment.new(elements)
       expected = {
-        "ISA" => {
-          "e01" => "00",
-          "e02" => "00",
+        "B2" => {
+          "e01" => "",
+          "e02" => "SCAC",
           "e03" => "",
-          "e04" => "09",
-          "e05" => "005070479ff",
-          "e06" => "ZZ",
-          "e07" => "U",
-          "e08" => "0",
-          "e09" => "T",
-          "e10" => "^"
+          "e04" => "124592j3",
+          "e05" => "",
+          "e06" => "CC"
         }
       }
       expect(segment.to_h).to eq(expected)
